% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_exists.R
\name{cache_exists}
\alias{cache_exists}
\title{cache_exists
checks if given path exists in the cache directory}
\usage{
cache_exists(level, path)
}
\arguments{
\item{level}{result rollup level (one of: psms, precursors, peptides, modified_peptides, protein_groups)}

\item{path}{s3 path to the parquet file}
}
\value{
True or False
}
\description{
cache_exists
checks if given path exists in the cache directory
}
\examples{
\dontrun{
  cache_exists("")
}
}
