% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cache.R
\name{clean_cache}
\alias{clean_cache}
\title{clean_cache
cleans the caches of given experiment names and levels}
\usage{
clean_cache(experiment_names = NULL, level = NULL, delete_all = F)
}
\arguments{
\item{experiment_names}{character vector of experiment names}

\item{level}{psms,precursors,protein_groups}

\item{delete_all}{boolean that deletes all of the caches}
}
\value{
none
}
\description{
clean_cache
cleans the caches of given experiment names and levels
}
\examples{
 \dontrun{
  clean_cache(c("exp1", "exp2"), "psms")
}
}
