% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_filter.R
\name{concat_filter}
\alias{concat_filter}
\title{concat_filter}
\usage{
concat_filter(total_filter_str, partial_filter_str)
}
\arguments{
\item{total_filter_str}{total filter string}

\item{partial_filter_str}{partially added filter string}
}
\value{
result filter
}
\description{
concat_filter
}
\examples{
\dontrun{
concat_filter("","")
}
}
