% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter_experiment_list.R
\name{create_filter_experiment_list}
\alias{create_filter_experiment_list}
\title{create_filter_experiment_list}
\usage{
create_filter_experiment_list(
  name_includes = NULL,
  name = NULL,
  username = NULL,
  tags = c(),
  from = NULL,
  to = NULL
)
}
\arguments{
\item{name_includes}{if name includes certain sub-string}

\item{name}{exact experiment name}

\item{username}{the username filter}

\item{tags}{the tag of the experiments}

\item{from}{filter experiments created after this timestamp (ISO 8601 format or POSIXct object)}

\item{to}{filter experiments created before this timestamp (ISO 8601 format or POSIXct object)}
}
\value{
filter string
}
\description{
create_filter_experiment_list
}
\examples{
 \dontrun{
create_filter_experiment_list(name="hello")
}
}
