% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id_token.R
\name{create_id_token}
\alias{create_id_token}
\title{create_id_token
creates id token by providing refresh token to the coginito}
\usage{
create_id_token(refresh_token, cognito_token_url, cognito_client_id)
}
\arguments{
\item{refresh_token}{string}

\item{cognito_token_url}{string}

\item{cognito_client_id}{string}
}
\value{
tokens object
}
\description{
create_id_token
creates id token by providing refresh token to the coginito
}
\examples{
\dontrun{
  create_id_token("","","")
}
}
