% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_parquet_to_cache.R
\name{download_parquet_to_cache}
\alias{download_parquet_to_cache}
\title{download_parquet_to_cache
downloads the parquet file to the cache directory}
\usage{
download_parquet_to_cache(level, s3_path, presigned_url)
}
\arguments{
\item{level}{result rollup level (one of: psms, precursors, peptides, modified_peptides, protein_groups)}

\item{s3_path}{cache path}

\item{presigned_url}{presigned url for the experiment files}
}
\description{
download_parquet_to_cache
downloads the parquet file to the cache directory
}
\examples{
 \dontrun{
  create_cache("","","","","")
}
}
