% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_names_to_objects.R
\name{experiment_names_to_objects}
\alias{experiment_names_to_objects}
\title{experiment_names_to_objects
converts the experiment names to experiment objects}
\usage{
experiment_names_to_objects(experiment_names)
}
\arguments{
\item{experiment_names}{names of the experiments}
}
\value{
list of experiment objects (pageItems from API response)
}
\description{
experiment_names_to_objects
converts the experiment names to experiment objects
}
\examples{
\dontrun{
platform_read_parquets("")
}
}
