% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_var.R
\name{get_env_var}
\alias{get_env_var}
\title{get_env_var
gets environment variable}
\usage{
get_env_var(env_name)
}
\arguments{
\item{env_name}{environment variable name}
}
\value{
env variable value
}
\description{
get_env_var
gets environment variable
}
\examples{
 \dontrun{
 get_env_var("")
 }
}
