% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_token.R
\name{get_id_token}
\alias{get_id_token}
\title{get_id_token
returns the id token to the msaid platform}
\usage{
get_id_token()
}
\value{
id_token
}
\description{
get_id_token
returns the id token to the msaid platform
}
\examples{
\dontrun{
get_id_token()
}
}
