% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_internal_env.R
\name{init_internal_env}
\alias{init_internal_env}
\title{init_internal_env}
\usage{
init_internal_env()
}
\value{
nothing
}
\description{
init_internal_env
}
\examples{
\dontrun{
  init_internal_env()
}
}
