% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_read_parquets.R
\name{platform_read_parquets}
\alias{platform_read_parquets}
\title{platform_read_parquets
reads the downloaded experiment parquet file}
\usage{
platform_read_parquets(
  dir,
  level,
  max_q_value = 1,
  max_global_q_value = 1,
  include_decoys = F,
  include_columns = NULL,
  exclude_columns = c(),
  exclude_array_columns = F
)
}
\arguments{
\item{dir}{cache dir}

\item{level}{result rollup level (one of: psms, precursors, peptides, modified_peptides, protein_groups)}

\item{max_q_value}{maximum q value}

\item{max_global_q_value}{maximum global q value}

\item{include_decoys}{include decoys (TRUE/FALSE)}

\item{include_columns}{include columns (TRUE/FALSE)}

\item{exclude_columns}{exclude columns (TRUE/FALSE)}

\item{exclude_array_columns}{exclude array columns (TRUE/FALSE)}
}
\value{
data
}
\description{
platform_read_parquets
reads the downloaded experiment parquet file
}
\examples{
\dontrun{
platform_read_parquets("")
}
}
