% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_var.R
\docType{data}
\name{SESSION_VARS}
\alias{SESSION_VARS}
\title{Session variable names stored per-stage}
\format{
An object of class \code{character} of length 5.
}
\usage{
SESSION_VARS
}
\description{
Session variable names stored per-stage
}
\keyword{internal}
