% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env_var.R
\name{set_env_var}
\alias{set_env_var}
\title{set_env_var
sets environment variable}
\usage{
set_env_var(env_name, env_value)
}
\arguments{
\item{env_name}{environment variable name}

\item{env_value}{environment variable's value}
}
\description{
set_env_var
sets environment variable
}
\examples{
 \dontrun{
 set_env_var("","","")
 }
}
